Moved Entry and Dataseriesmapper out of IStoreData.ts
