+++ package.json
@@ -70,11 +70,13 @@
     "vuex-electron": "^1.0.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.1.7",
     "@types/encoding-down": "^5.0.0",
     "@types/leveldown": "^4.0.0",
     "@types/levelup": "^3.1.0",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.12.21",
+    "@types/rimraf": "^2.0.2",
     "ajv": "^6.5.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.3",
@@ -124,6 +126,7 @@
     "node-sass": "^4.9.2",
     "nyc": "^13.2.0",
     "require-dir": "^1.0.0",
+    "rimraf": "^2.6.3",
     "sass-loader": "^7.0.3",
     "source-map-support": "^0.5.10",
     "spectron": "^3.8.0",
@@ -133,6 +136,7 @@
     "tslint": "^5.12.1",
     "tslint-config-standard": "^8.0.1",
     "tslint-loader": "^3.5.4",
+    "tslint-no-unused-expression-chai": "^0.1.4",
     "typescript": "^3.2.4",
     "url-loader": "^1.0.1",
     "vue-html-loader": "^1.2.4",
@@ -161,7 +165,8 @@
       "**/*.d.ts"
     ],
     "reporter": [
-      "html"
+      "html",
+      "text-summary"
     ],
     "sourceMap": true,
     "report-dir": "./test/unit/backend/coverage",
